finalproject_fxalarm/requirements.txt
astroid==1.4.5
beautifulsoup4==4.4.1
colorama==0.3.7
Django==1.9.6
lazy-object-proxy==1.2.2
pylint==1.5.5
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
wrapt==1.10.8
